requirements.txt
django-admin-sortable2
django-dynamic-preferences
reportlab